Refactored stripe checkout creation
